requirements.txt
alembic==1.7.1
aniso8601==9.0.1
atomicwrites==1.4.1
attrs==20.3.0
bcrypt==3.2.0
blinker==1.4
certifi==2021.5.30
cffi==1.14.4
chardet==4.0.0
click==7.1.2
cmake==3.22.5
colorama==0.4.4
compreface-sdk==0.6.0
conekta==2.6.0
coverage==5.4
cryptography==3.3.1
cx-Freeze==6.6
cx-Logging==3.0
cycler==0.10.0
dnspython==2.2.1
ecdsa==0.17.0
email-validator==1.2.1
face-recognition-models==0.3.0
Flask==1.1.2
Flask-Cors==3.0.10
Flask-JWT==0.3.2
Flask-Login==0.5.0
Flask-Mail==0.9.1
flask-marshmallow==0.14.0
Flask-Migrate==3.1.0
Flask-QR==0.1.3
Flask-QRcode==3.0.0
Flask-ReCaptcha==0.4.2
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.4.4
 Flask-JWT==0.3.2
Flask-Testing==0.8.1
Flask-JWT-Extended==4.4.4
Flask-WTF==0.14.3
future==0.15.2
greenlet==1.1.1
gunicorn==20.0.4
h5py==3.7.0
idna==2.10
importlib-metadata==3.4.0
iniconfig==1.1.1
itsdangerous==1.1.0
Jinja2==2.11.2
kiwisolver==1.3.1
Mako==1.1.5
MarkupSafe==1.1.1
marshmallow==3.17.1
matplotlib==3.4.2
nose==1.3.7
numpy==1.21.0
opencv-python==4.6.0.66
openpay==0.6
packaging==20.9
pandas==1.3.3
paramiko==2.7.2
passlib==1.7.4
Pillow==8.1.0
pkipplib==0.7
pluggy==0.13.1
psycopg2-binary==2.9.1
py==1.10.0
pycparser==2.20
pycryptodome==3.3.1
PyJWT==1.4.2
PyMySQL==1.0.2
PyNaCl==1.4.0
pyparsing==2.4.7
pypng==0.0.21
PyQRCode==1.2.1
pyserial==3.5
pytest==6.2.2
pytest-cov==2.11.1
python-dateutil==2.8.1
python-jose-cryptodome==1.3.2
pytz==2021.1
pyzbar==0.1.8
qrcode==6.1
requests==2.25.1
requests-toolbelt==0.9.1
simplejson==3.17.2
six==1.15.0
SQLAlchemy==1.3.22
sshtunnel==0.4.0
toml==0.10.2
typing-extensions==3.7.4.3
urllib3==1.26.6
Werkzeug==1.0.1
WTForms==2.3.3
zipp==3.4.0
